+++ packages/filterlist/CHANGELOG.md
@@ -4,6 +4,8 @@
 
 * Added `page` to the state of list
 * Added `setPage` method
+* Added `total` to the state of list
+* Added `setTotal` method
 
 ## 0.3.1 (04 sep 2023)